fix: remove client metadata exports
