+++ frontend/src/app/assistant/page.tsx
@@ -1,4 +1,3 @@
-'use client';
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
